Support decimal values for Annual Paid Days (#12) (#13)
* Support decimal values for Annual Paid Days (PTO allowance)

Changes int to decimal(18,2) across the model, ViewModel, balance
calculation, and input so consultants can be assigned fractional
PTO days (e.g. 15.5). Also gitignores launchSettings.json and
CLAUDE.md to prevent local dev secrets from being committed.

* Allow any decimal value for Annual Paid Days input

---------

+++ OceansApp/OceansApp.Models/ViewModels/TimeOff/TimeOffBalancesVM.cs
@@ -3,8 +3,8 @@ namespace OceansApp.Models.ViewModels.TimeOff
     public class TimeOffBalancesVM
     {
         public bool IsPtoEnabled { get; set; }
-        public int PtoAvailable { get; set; }
-        public int PtoAnnualAllowance { get; set; }
+        public decimal PtoAvailable { get; set; }
+        public decimal PtoAnnualAllowance { get; set; }
         public int VtoAvailable { get; set; }
     }
 }
